requirements.txt
fastapi[standard]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
python-jose[cryptography]
redis
httpx
python-multipart
argon2-cffi
authx
pwdlib
greenlet
lark
pytest
pytest-asyncio
httpx